Add ModelForNPMC, DetectPostPart
ModelForNPMC
We define a new Model class that is removed 'augment', 'profile' conditional statements to fix the control flow.

DetectPostPart
Post part of the detection that is only used in test mode is detached when the model is exported to fx.GraphModule format.
To address this problem, we define DetectPostPart that functionate the detached part.